+++ omniva-woocommerce/core/methods/class-method-postspecific.php
@@ -8,6 +8,9 @@ class OmnivaLt_Method_PostSpecific extends OmnivaLt_Method_Core
         $this->setTitle(__('Specific post office', 'omnivalt'));
         $this->setDescription(__('Activate this service, when you want to send parcels to specific post office.', 'omnivalt'));
         $this->setIsShippingMethod(true);
-        $this->setDefaultWeight(100);
+        $this->setTerminalsType('post');
+        $this->setMaxWeight(100);
+        $this->setRestrictApi(array('EE'));
+        $this->setRestrictCountry(array('EE'));
     }
 }